Fix MMSI bug
The code had an error in how it unpacked the mmsi from the web response.

Document more of the web response values.

Document deprecated fields.

AIS is no longer reported as an int. Use the string form instead. This
means that the uninterestingAIS table will likely have to be rebuilt by
trial and error.