Tuning
Eliminate a few more AIS vessel types.

Don't fast loop on failures to communicate with REST APIs. It runs the CPU in a hot loop when there is no Internet.

Remove "found ship" logging so that errors don't scroll off the screen so quickly. We want to see the errors.