Some ships we always want to see
The Matthew Turner, for example, is an interesting ship and we want to be notified if she is in sight. But, she is typed as 'Passenger Ship' and we have blocklisted that type. So, add an override interesting list that will allow her (and any others we add) to be seen.