Refactor main to its own dir
It turns out that one cannot have multiple files that are all part of
package main. There can only be one file in the main package. Move main
into its own directory. Rename the main package to ship_ahoy. That has
its own problem, as one should not have an underscore in a module name.
But, renaming the entire repo will have to be a job for a later commit.

Also in a later commit, build a better entry point into the shipahoy
module than the `Doit()` function. :)